Get Project compiling after LspStore extraction
