+++ crates/remote_server/src/headless_project.rs
@@ -41,9 +41,12 @@ impl HeadlessProject {
     pub fn new(session: Arc<SshSession>, fs: Arc<dyn Fs>, cx: &mut ModelContext<Self>) -> Self {
         let this = cx.weak_model();
 
-        let worktree_store = cx.new_model(|_| WorktreeStore::new(true));
-        let buffer_store =
-            cx.new_model(|cx| BufferStore::new(worktree_store.clone(), Some(PROJECT_ID), cx));
+        let worktree_store = cx.new_model(|_| WorktreeStore::new(true, fs.clone()));
+        let buffer_store = cx.new_model(|cx| {
+            let mut buffer_store = BufferStore::new(worktree_store.clone(), Some(PROJECT_ID), cx);
+            buffer_store.shared(PROJECT_ID, session.clone().into(), cx);
+            buffer_store
+        });
         cx.subscribe(&buffer_store, Self::on_buffer_store_event)
             .detach();
 
@@ -128,7 +131,7 @@ impl HeadlessProject {
         mut cx: AsyncAppContext,
     ) -> Result<proto::OpenBufferResponse> {
         let worktree_id = WorktreeId::from_proto(message.payload.worktree_id);
-        let (buffer_store, buffer, session) = this.update(&mut cx, |this, cx| {
+        let (buffer_store, buffer) = this.update(&mut cx, |this, cx| {
             let buffer_store = this.buffer_store.clone();
             let buffer = this.buffer_store.update(cx, |buffer_store, cx| {
                 buffer_store.open_buffer(
@@ -139,14 +142,14 @@ impl HeadlessProject {
                     cx,
                 )
             });
-            anyhow::Ok((buffer_store, buffer, this.session.clone()))
+            anyhow::Ok((buffer_store, buffer))
         })??;
 
         let buffer = buffer.await?;
         let buffer_id = buffer.read_with(&cx, |b, _| b.remote_id())?;
         buffer_store.update(&mut cx, |buffer_store, cx| {
             buffer_store
-                .create_buffer_for_peer(&buffer, PEER_ID, PROJECT_ID, session, cx)
+                .create_buffer_for_peer(&buffer, PEER_ID, cx)
                 .detach_and_log_err(cx);
         })?;
 
@@ -176,22 +179,14 @@ impl HeadlessProject {
             buffer_ids: Vec::new(),
         };
 
-        let (buffer_store, client) = this.update(&mut cx, |this, _| {
-            (this.buffer_store.clone(), this.session.clone())
-        })?;
+        let buffer_store = this.read_with(&cx, |this, _| this.buffer_store.clone())?;
 
         while let Some(buffer) = results.next().await {
             let buffer_id = buffer.update(&mut cx, |this, _| this.remote_id())?;
             response.buffer_ids.push(buffer_id.to_proto());
             buffer_store
                 .update(&mut cx, |buffer_store, cx| {
-                    buffer_store.create_buffer_for_peer(
-                        &buffer,
-                        PEER_ID,
-                        PROJECT_ID,
-                        client.clone(),
-                        cx,
-                    )
+                    buffer_store.create_buffer_for_peer(&buffer, PEER_ID, cx)
                 })?
                 .await?;
         }
